Implementa um endpoint para recuperar a avaliação de uma trilha do usuário autenticado
